+++ packages/theme/src/unified/UnifiedThemeProvider.tsx
@@ -43,14 +43,14 @@ export interface UnifiedThemeProviderProps {
 
 // Background at https://mui.com/x/migration/migration-data-grid-v4/#using-mui-core-v4-with-v5
 // Rather than disabling globals and custom seed, we instead only set a production prefix that
-// won't collide with MUI 5 styles. We've already got a separate class name generator for v5 set
+// won't collide with Material UI 5 styles. We've already got a separate class name generator for v5 set
 // up in MuiClassNameSetup.ts, so only the production JSS needs deduplication.
 const generateV4ClassName = createGenerateClassName({
   productionPrefix: 'jss4-',
 });
 
 /**
- * Provides themes for all MUI versions supported by the provided unified theme.
+ * Provides themes for all Material UI versions supported by the provided unified theme.
  *
  * @public
  */