+++ packages/backend-next/src/index.ts
@@ -37,6 +37,7 @@ import { azureDevOpsPlugin } from '@backstage/plugin-azure-devops-backend';
 import { linguistPlugin } from '@backstage/plugin-linguist-backend';
 import { devtoolsPlugin } from '@backstage/plugin-devtools-backend';
 import { TaskScheduleDefinition } from '@backstage/backend-tasks';
+import { adrPlugin } from '@backstage/plugin-adr-backend';
 
 const backend = createBackend();
 
@@ -72,6 +73,8 @@ backend.add(
 
 // Todo
 backend.add(todoPlugin());
+
+backend.add(adrPlugin());
 
 // Techdocs
 backend.add(techdocsPlugin());