+++ plugins/catalog/src/components/DependsOnResourcesCard/DependsOnResourcesCard.tsx
@@ -15,7 +15,11 @@
  */
 
 import { RELATION_DEPENDS_ON, ResourceEntity } from '@backstage/catalog-model';
-import { InfoCardVariants, TableColumn } from '@backstage/core-components';
+import {
+  InfoCardVariants,
+  TableColumn,
+  TableOptions,
+} from '@backstage/core-components';
 import React from 'react';
 import {
   asResourceEntities,
@@ -29,6 +33,7 @@ export interface DependsOnResourcesCardProps {
   variant?: InfoCardVariants;
   title?: string;
   columns?: TableColumn<ResourceEntity>[];
+  tableOptions?: TableOptions;
 }
 
 export function DependsOnResourcesCard(props: DependsOnResourcesCardProps) {
@@ -36,6 +41,7 @@ export function DependsOnResourcesCard(props: DependsOnResourcesCardProps) {
     variant = 'gridItem',
     title = 'Depends on resources',
     columns = resourceEntityColumns,
+    tableOptions = {},
   } = props;
   return (
     <RelatedEntitiesCard
@@ -47,6 +53,7 @@ export function DependsOnResourcesCard(props: DependsOnResourcesCardProps) {
       emptyMessage="No resource is a dependency of this component"
       emptyHelpLink={componentEntityHelpLink}
       asRenderableEntities={asResourceEntities}
+      tableOptions={tableOptions}
     />
   );
 }